Refatorar localizador por GPS
